Refactor: Changed Java version to 17 and changed directory structure and renaming the directory
