+++ src/main/java/com/example/HomeSaveHome/BulletinBoard/api/ArtitcleAPIController.java
@@ -1,9 +1,9 @@
-package com.example.demo.api;
+package com.example.HomeSaveHome.BulletinBoard.api;
 
-import com.example.demo.dto.ArticleForm;
-import com.example.demo.entity.Article;
-import com.example.demo.repository.ArticleRepository;
-import com.example.demo.service.ArticleService;
+import com.example.HomeSaveHome.BulletinBoard.dto.ArticleForm;
+import com.example.HomeSaveHome.BulletinBoard.entity.Article;
+import com.example.HomeSaveHome.BulletinBoard.repository.ArticleRepository;
+import com.example.HomeSaveHome.BulletinBoard.service.ArticleService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
