+++ src/main/java/com/example/HomeSaveHome/BulletinBoard/api/FirstAPIController.java
@@ -1,4 +1,4 @@
-package com.example.demo.api;
+package com.example.HomeSaveHome.BulletinBoard.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
