+++ src/main/java/com/example/HomeSaveHome/BulletinBoard/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo.config;
+package com.example.HomeSaveHome.BulletinBoard.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
