+++ src/main/java/com/example/HomeSaveHome/BulletinBoard/controller/FirstController.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.HomeSaveHome.BulletinBoard.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
