+++ src/main/java/com/example/HomeSaveHome/HomeSaveHomeApplication.java
@@ -1,13 +1,13 @@
-package com.example.demo;
+package com.example.HomeSaveHome;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Demo2Application {
+public class HomeSaveHomeApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Demo2Application.class, args);
+        SpringApplication.run(HomeSaveHomeApplication.class, args);
     }
 
 }
