--- src/test/java/com/example/demo/Demo2ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Demo2ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
